requirement.txt
asgiref==3.4.1
crispy-tailwind==0.5.0
Django==3.2.7
django-countries==7.2.1
django-crispy-forms==1.12.0
django-environ==0.6.0
django-phonenumber-field==5.2.0
gunicorn==20.1.0
phonenumbers==8.12.31
Pillow==8.3.2
pytz==2021.1
pyuca==1.2
sqlparse==0.4.1
whitenoise==5.3.0
